Network.AWS.CloudHSM.CreateHAPG

Description

Creates a high-availability partition group. A high-availability partition group is a group of partitions that spans multiple physical HSMs.

Creating a Request

createHAPG

Arguments

:: Text
chLabel

-> CreateHAPG

Creates a value of CreateHAPG with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chLabel - The label of the new high-availability partition group.

data CreateHAPG

Contains the inputs for the CreateHapgRequest action.

See: createHAPG smart constructor.

Request Lenses

chLabel :: Lens' CreateHAPG Text

The label of the new high-availability partition group.

Destructuring the Response

createHAPGResponse

Arguments

:: Int
chapgrsResponseStatus

-> CreateHAPGResponse

Creates a value of CreateHAPGResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chapgrsHAPGARN - The ARN of the high-availability partition group.

chapgrsResponseStatus - -- | The response status code.

data CreateHAPGResponse

Contains the output of the CreateHAPartitionGroup action.

See: createHAPGResponse smart constructor.

Response Lenses

chapgrsHAPGARN :: Lens' CreateHAPGResponse (Maybe Text)

The ARN of the high-availability partition group.

chapgrsResponseStatus :: Lens' CreateHAPGResponse Int

- | The response status code.