Network.AWS.CloudHSM.CreateHSM

Description

Creates an uninitialized HSM instance.

There is an upfront fee charged for each HSM instance that you create with the CreateHsm operation. If you accidentally provision an HSM and want to request a refund, delete the instance using the DeleteHsm operation, go to the AWS Support Center , create a new case, and select Account and Billing Support .

Important: It can take up to 20 minutes to create and provision an HSM. You can monitor the status of the HSM with the DescribeHsm operation. The HSM is ready to be initialized when the status changes to RUNNING .

Creating a Request

createHSM

Arguments

:: Text
chSubnetId

-> Text
chSSHKey

-> Text
chIAMRoleARN

-> SubscriptionType
chSubscriptionType

-> CreateHSM

Creates a value of CreateHSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chClientToken - A user-defined token to ensure idempotence. Subsequent calls to this operation with the same token will be ignored.

chSyslogIP - The IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

chExternalId - The external ID from IamRoleArn , if present.

chEniIP - The IP address to assign to the HSM's ENI. If an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the subnet.

chSubnetId - The identifier of the subnet in your VPC in which to place the HSM.

chSSHKey - The SSH public key to install on the HSM.

chIAMRoleARN - The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

chSubscriptionType - Undocumented member.

data CreateHSM

Contains the inputs for the CreateHsm operation.

See: createHSM smart constructor.

Request Lenses

chClientToken :: Lens' CreateHSM (Maybe Text)

A user-defined token to ensure idempotence. Subsequent calls to this operation with the same token will be ignored.

chSyslogIP :: Lens' CreateHSM (Maybe Text)

The IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

chExternalId :: Lens' CreateHSM (Maybe Text)

The external ID from IamRoleArn , if present.

chEniIP :: Lens' CreateHSM (Maybe Text)

The IP address to assign to the HSM's ENI. If an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the subnet.

chSubnetId :: Lens' CreateHSM Text

The identifier of the subnet in your VPC in which to place the HSM.

chSSHKey :: Lens' CreateHSM Text

The SSH public key to install on the HSM.

chIAMRoleARN :: Lens' CreateHSM Text

The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

chSubscriptionType :: Lens' CreateHSM SubscriptionType

Undocumented member.

Destructuring the Response

createHSMResponse

Arguments

:: Int
chrsResponseStatus

-> CreateHSMResponse

Creates a value of CreateHSMResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chrsHSMARN - The ARN of the HSM.

chrsResponseStatus - -- | The response status code.

data CreateHSMResponse

Contains the output of the CreateHsm operation.

See: createHSMResponse smart constructor.

Response Lenses

chrsHSMARN :: Lens' CreateHSMResponse (Maybe Text)

The ARN of the HSM.

chrsResponseStatus :: Lens' CreateHSMResponse Int

- | The response status code.