Network.AWS.CloudHSM.CreateHsm

Description

Creates an uninitialized HSM instance. Running this command provisions an HSM appliance and will result in charges to your AWS account for the HSM.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_CreateHsm.html

Request

data CreateHsm

Request constructor

createHsm

Arguments

:: Text
chSubnetId

-> Text
chSshKey

-> Text
chIamRoleArn

-> SubscriptionType
chSubscriptionType

-> CreateHsm

CreateHsm constructor.

The fields accessible through corresponding lenses are:

chClientToken :: Maybe Text

chEniIp :: Maybe Text

chExternalId :: Maybe Text

chIamRoleArn :: Text

chSshKey :: Text

chSubnetId :: Text

chSubscriptionType :: SubscriptionType

chSyslogIp :: Maybe Text

Request lenses

chClientToken :: Lens' CreateHsm (Maybe Text)

A user-defined token to ensure idempotence. Subsequent calls to this action with the same token will be ignored.

chEniIp :: Lens' CreateHsm (Maybe Text)

The IP address to assign to the HSM's ENI.

chExternalId :: Lens' CreateHsm (Maybe Text)

The external ID from IamRoleArn, if present.

chIamRoleArn :: Lens' CreateHsm Text

The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

chSshKey :: Lens' CreateHsm Text

The SSH public key to install on the HSM.

chSubnetId :: Lens' CreateHsm Text

The identifier of the subnet in your VPC in which to place the HSM.

chSubscriptionType :: Lens' CreateHsm SubscriptionType

The subscription type.

chSyslogIp :: Lens' CreateHsm (Maybe Text)

The IP address for the syslog monitoring server.

Response

data CreateHsmResponse

Response constructor

createHsmResponse :: CreateHsmResponse

CreateHsmResponse constructor.

The fields accessible through corresponding lenses are:

chrHsmArn :: Maybe Text

Response lenses

chrHsmArn :: Lens' CreateHsmResponse (Maybe Text)

The ARN of the HSM.