Network.AWS.CloudHSM.CreateLunaClient

Description

Creates an HSM client.

Creating a Request

createLunaClient

Arguments

:: Text
clcCertificate

-> CreateLunaClient

Creates a value of CreateLunaClient with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clcLabel - The label for the client.

clcCertificate - The contents of a Base64-Encoded X.509 v3 certificate to be installed on the HSMs used by this client.

data CreateLunaClient

Contains the inputs for the CreateLunaClient action.

See: createLunaClient smart constructor.

Request Lenses

clcLabel :: Lens' CreateLunaClient (Maybe Text)

The label for the client.

clcCertificate :: Lens' CreateLunaClient Text

The contents of a Base64-Encoded X.509 v3 certificate to be installed on the HSMs used by this client.

Destructuring the Response

createLunaClientResponse

Arguments

:: Int
clcrsResponseStatus

-> CreateLunaClientResponse

Creates a value of CreateLunaClientResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clcrsClientARN - The ARN of the client.

clcrsResponseStatus - -- | The response status code.

data CreateLunaClientResponse

Contains the output of the CreateLunaClient action.

See: createLunaClientResponse smart constructor.

Response Lenses

clcrsClientARN :: Lens' CreateLunaClientResponse (Maybe Text)

The ARN of the client.

clcrsResponseStatus :: Lens' CreateLunaClientResponse Int

- | The response status code.