Network.AWS.CloudHSM.DeleteHSM

Description

Deletes an HSM. After completion, this operation cannot be undone and your key material cannot be recovered.

Creating a Request

deleteHSM

Arguments

:: Text
dhHSMARN

-> DeleteHSM

Creates a value of DeleteHSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhHSMARN - The ARN of the HSM to delete.

data DeleteHSM

Contains the inputs for the DeleteHsm operation.

See: deleteHSM smart constructor.

Request Lenses

dhHSMARN :: Lens' DeleteHSM Text

The ARN of the HSM to delete.

Destructuring the Response

deleteHSMResponse

Arguments

:: Int
dhsmrsResponseStatus

-> Text
dhsmrsStatus

-> DeleteHSMResponse

Creates a value of DeleteHSMResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhsmrsResponseStatus - -- | The response status code.

dhsmrsStatus - The status of the operation.

data DeleteHSMResponse

Contains the output of the DeleteHsm operation.

See: deleteHSMResponse smart constructor.

Response Lenses

dhsmrsResponseStatus :: Lens' DeleteHSMResponse Int

- | The response status code.

dhsmrsStatus :: Lens' DeleteHSMResponse Text

The status of the operation.