Network.AWS.CloudHSM.DescribeHAPG

Description

Retrieves information about a high-availability partition group.

Creating a Request

describeHAPG

Arguments

:: Text
dhapgHAPGARN

-> DescribeHAPG

Creates a value of DescribeHAPG with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhapgHAPGARN - The ARN of the high-availability partition group to describe.

data DescribeHAPG

Contains the inputs for the DescribeHapg action.

See: describeHAPG smart constructor.

Request Lenses

dhapgHAPGARN :: Lens' DescribeHAPG Text

The ARN of the high-availability partition group to describe.

Destructuring the Response

describeHAPGResponse

Arguments

:: Int
dhapgrsResponseStatus

-> DescribeHAPGResponse

Creates a value of DescribeHAPGResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhapgrsState - The state of the high-availability partition group.

dhapgrsLastModifiedTimestamp - The date and time the high-availability partition group was last modified.

dhapgrsHSMsPendingRegistration - Undocumented member.

dhapgrsHSMsPendingDeletion - Undocumented member.

dhapgrsHAPGSerial - The serial number of the high-availability partition group.

dhapgrsHSMsLastActionFailed - Undocumented member.

dhapgrsPartitionSerialList - The list of partition serial numbers that belong to the high-availability partition group.

dhapgrsHAPGARN - The ARN of the high-availability partition group.

dhapgrsLabel - The label for the high-availability partition group.

dhapgrsResponseStatus - -- | The response status code.

data DescribeHAPGResponse

Contains the output of the DescribeHapg action.

See: describeHAPGResponse smart constructor.

Response Lenses

dhapgrsState :: Lens' DescribeHAPGResponse (Maybe CloudHSMObjectState)

The state of the high-availability partition group.

dhapgrsLastModifiedTimestamp :: Lens' DescribeHAPGResponse (Maybe Text)

The date and time the high-availability partition group was last modified.

dhapgrsHSMsPendingRegistration :: Lens' DescribeHAPGResponse [Text]

Undocumented member.

dhapgrsHSMsPendingDeletion :: Lens' DescribeHAPGResponse [Text]

Undocumented member.

dhapgrsHAPGSerial :: Lens' DescribeHAPGResponse (Maybe Text)

The serial number of the high-availability partition group.

dhapgrsHSMsLastActionFailed :: Lens' DescribeHAPGResponse [Text]

Undocumented member.

dhapgrsPartitionSerialList :: Lens' DescribeHAPGResponse [Text]

The list of partition serial numbers that belong to the high-availability partition group.

dhapgrsHAPGARN :: Lens' DescribeHAPGResponse (Maybe Text)

The ARN of the high-availability partition group.

dhapgrsLabel :: Lens' DescribeHAPGResponse (Maybe Text)

The label for the high-availability partition group.

dhapgrsResponseStatus :: Lens' DescribeHAPGResponse Int

- | The response status code.