Network.AWS.CloudHSM.DescribeHSM

Description

Retrieves information about an HSM. You can identify the HSM by its ARN or its serial number.

Creating a Request

describeHSM :: DescribeHSM

Creates a value of DescribeHSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhsmHSMSerialNumber - The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

dhsmHSMARN - The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

data DescribeHSM

Contains the inputs for the DescribeHsm operation.

See: describeHSM smart constructor.

Request Lenses

dhsmHSMSerialNumber :: Lens' DescribeHSM (Maybe Text)

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

dhsmHSMARN :: Lens' DescribeHSM (Maybe Text)

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

Destructuring the Response

describeHSMResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeHSMResponse

Creates a value of DescribeHSMResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsStatus - The status of the HSM.

desrsIAMRoleARN - The ARN of the IAM role assigned to the HSM.

desrsEniId - The identifier of the elastic network interface (ENI) attached to the HSM.

desrsVPCId - The identifier of the VPC that the HSM is in.

desrsSSHKeyLastUpdated - The date and time that the SSH key was last updated.

desrsSubscriptionEndDate - The subscription end date.

desrsServerCertURI - The URI of the certificate server.

desrsSubscriptionType - Undocumented member.

desrsSSHPublicKey - The public SSH key.

desrsSubnetId - The identifier of the subnet that the HSM is in.

desrsStatusDetails - Contains additional information about the status of the HSM.

desrsPartitions - The list of partitions on the HSM.

desrsSubscriptionStartDate - The subscription start date.

desrsAvailabilityZone - The Availability Zone that the HSM is in.

desrsServerCertLastUpdated - The date and time that the server certificate was last updated.

desrsSoftwareVersion - The HSM software version.

desrsVendorName - The name of the HSM vendor.

desrsSerialNumber - The serial number of the HSM.

desrsHSMARN - The ARN of the HSM.

desrsEniIP - The IP address assigned to the HSM's ENI.

desrsHSMType - The HSM model type.

desrsResponseStatus - -- | The response status code.

data DescribeHSMResponse

Contains the output of the DescribeHsm operation.

See: describeHSMResponse smart constructor.

Response Lenses

desrsStatus :: Lens' DescribeHSMResponse (Maybe HSMStatus)

The status of the HSM.

desrsIAMRoleARN :: Lens' DescribeHSMResponse (Maybe Text)

The ARN of the IAM role assigned to the HSM.

desrsEniId :: Lens' DescribeHSMResponse (Maybe Text)

The identifier of the elastic network interface (ENI) attached to the HSM.

desrsVPCId :: Lens' DescribeHSMResponse (Maybe Text)

The identifier of the VPC that the HSM is in.

desrsSSHKeyLastUpdated :: Lens' DescribeHSMResponse (Maybe Text)

The date and time that the SSH key was last updated.

desrsSubscriptionEndDate :: Lens' DescribeHSMResponse (Maybe Text)

The subscription end date.

desrsServerCertURI :: Lens' DescribeHSMResponse (Maybe Text)

The URI of the certificate server.

desrsSubscriptionType :: Lens' DescribeHSMResponse (Maybe SubscriptionType)

Undocumented member.

desrsSSHPublicKey :: Lens' DescribeHSMResponse (Maybe Text)

The public SSH key.

desrsSubnetId :: Lens' DescribeHSMResponse (Maybe Text)

The identifier of the subnet that the HSM is in.

desrsStatusDetails :: Lens' DescribeHSMResponse (Maybe Text)

Contains additional information about the status of the HSM.

desrsPartitions :: Lens' DescribeHSMResponse [Text]

The list of partitions on the HSM.

desrsSubscriptionStartDate :: Lens' DescribeHSMResponse (Maybe Text)

The subscription start date.

desrsAvailabilityZone :: Lens' DescribeHSMResponse (Maybe Text)

The Availability Zone that the HSM is in.

desrsServerCertLastUpdated :: Lens' DescribeHSMResponse (Maybe Text)

The date and time that the server certificate was last updated.

desrsSoftwareVersion :: Lens' DescribeHSMResponse (Maybe Text)

The HSM software version.

desrsVendorName :: Lens' DescribeHSMResponse (Maybe Text)

The name of the HSM vendor.

desrsSerialNumber :: Lens' DescribeHSMResponse (Maybe Text)

The serial number of the HSM.

desrsHSMARN :: Lens' DescribeHSMResponse (Maybe Text)

The ARN of the HSM.

desrsEniIP :: Lens' DescribeHSMResponse (Maybe Text)

The IP address assigned to the HSM's ENI.

desrsHSMType :: Lens' DescribeHSMResponse (Maybe Text)

The HSM model type.

desrsResponseStatus :: Lens' DescribeHSMResponse Int

- | The response status code.