Network.AWS.CloudHSM.DescribeHsm

Description

Retrieves information about an HSM. You can identify the HSM by its ARN or its serial number.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_DescribeHsm.html

Request

data DescribeHsm

Request constructor

describeHsm :: DescribeHsm

DescribeHsm constructor.

The fields accessible through corresponding lenses are:

dh1HsmArn :: Maybe Text

dh1HsmSerialNumber :: Maybe Text

Request lenses

dh1HsmArn :: Lens' DescribeHsm (Maybe Text)

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

dh1HsmSerialNumber :: Lens' DescribeHsm (Maybe Text)

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

Response

data DescribeHsmResponse

Response constructor

describeHsmResponse :: DescribeHsmResponse

DescribeHsmResponse constructor.

The fields accessible through corresponding lenses are:

dhr2AvailabilityZone :: Maybe Text

dhr2EniId :: Maybe Text

dhr2EniIp :: Maybe Text

dhr2HsmArn :: Maybe Text

dhr2HsmType :: Maybe Text

dhr2IamRoleArn :: Maybe Text

dhr2Partitions :: [Text]

dhr2SerialNumber :: Maybe Text

dhr2ServerCertLastUpdated :: Maybe Text

dhr2ServerCertUri :: Maybe Text

dhr2SoftwareVersion :: Maybe Text

dhr2SshKeyLastUpdated :: Maybe Text

dhr2SshPublicKey :: Maybe Text

dhr2Status :: Maybe HsmStatus

dhr2StatusDetails :: Maybe Text

dhr2SubnetId :: Maybe Text

dhr2SubscriptionEndDate :: Maybe Text

dhr2SubscriptionStartDate :: Maybe Text

dhr2SubscriptionType :: Maybe SubscriptionType

dhr2VendorName :: Maybe Text

dhr2VpcId :: Maybe Text

Response lenses

dhr2AvailabilityZone :: Lens' DescribeHsmResponse (Maybe Text)

The Availability Zone that the HSM is in.

dhr2EniId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the elastic network interface (ENI) attached to the HSM.

dhr2EniIp :: Lens' DescribeHsmResponse (Maybe Text)

The IP address assigned to the HSM's ENI.

dhr2HsmArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the HSM.

dhr2HsmType :: Lens' DescribeHsmResponse (Maybe Text)

The HSM model type.

dhr2IamRoleArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the IAM role assigned to the HSM.

dhr2Partitions :: Lens' DescribeHsmResponse [Text]

The list of partitions on the HSM.

dhr2SerialNumber :: Lens' DescribeHsmResponse (Maybe Text)

The serial number of the HSM.

dhr2ServerCertLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time the server certificate was last updated.

dhr2ServerCertUri :: Lens' DescribeHsmResponse (Maybe Text)

The URI of the certificate server.

dhr2SoftwareVersion :: Lens' DescribeHsmResponse (Maybe Text)

The HSM software version.

dhr2SshKeyLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time the SSH key was last updated.

dhr2SshPublicKey :: Lens' DescribeHsmResponse (Maybe Text)

The public SSH key.

dhr2Status :: Lens' DescribeHsmResponse (Maybe HsmStatus)

The status of the HSM.

dhr2StatusDetails :: Lens' DescribeHsmResponse (Maybe Text)

Contains additional information about the status of the HSM.

dhr2SubnetId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the subnet the HSM is in.

dhr2SubscriptionEndDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription end date.

dhr2SubscriptionStartDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription start date.

dhr2SubscriptionType :: Lens' DescribeHsmResponse (Maybe SubscriptionType)

The subscription type.

dhr2VendorName :: Lens' DescribeHsmResponse (Maybe Text)

The name of the HSM vendor.

dhr2VpcId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the VPC that the HSM is in.