Network.AWS.CloudHSM.DescribeLunaClient

Description

Retrieves information about an HSM client.

Creating a Request

describeLunaClient :: DescribeLunaClient

Creates a value of DescribeLunaClient with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlcClientARN - The ARN of the client.

dlcCertificateFingerprint - The certificate fingerprint.

data DescribeLunaClient

See: describeLunaClient smart constructor.

Request Lenses

dlcClientARN :: Lens' DescribeLunaClient (Maybe Text)

The ARN of the client.

dlcCertificateFingerprint :: Lens' DescribeLunaClient (Maybe Text)

The certificate fingerprint.

Destructuring the Response

describeLunaClientResponse

Arguments

:: Int
drsResponseStatus

-> DescribeLunaClientResponse

Creates a value of DescribeLunaClientResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsClientARN - The ARN of the client.

drsLastModifiedTimestamp - The date and time the client was last modified.

drsCertificateFingerprint - The certificate fingerprint.

drsCertificate - The certificate installed on the HSMs used by this client.

drsLabel - The label of the client.

drsResponseStatus - -- | The response status code.

data DescribeLunaClientResponse

See: describeLunaClientResponse smart constructor.

Response Lenses

drsClientARN :: Lens' DescribeLunaClientResponse (Maybe Text)

The ARN of the client.

drsLastModifiedTimestamp :: Lens' DescribeLunaClientResponse (Maybe Text)

The date and time the client was last modified.

drsCertificateFingerprint :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate fingerprint.

drsCertificate :: Lens' DescribeLunaClientResponse (Maybe Text)

The certificate installed on the HSMs used by this client.

drsLabel :: Lens' DescribeLunaClientResponse (Maybe Text)

The label of the client.

drsResponseStatus :: Lens' DescribeLunaClientResponse Int

- | The response status code.