Network.AWS.CloudHSM.ListHapgs

Description

Lists the high-availability partition groups for the account.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHapgs to retrieve the next set of items.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_ListHapgs.html

Request

data ListHapgs

Request constructor

listHapgs :: ListHapgs

ListHapgs constructor.

The fields accessible through corresponding lenses are:

lhNextToken :: Maybe Text

Request lenses

lhNextToken :: Lens' ListHapgs (Maybe Text)

The NextToken value from a previous call to ListHapgs. Pass null if this is the first call.

Response

data ListHapgsResponse

Response constructor

listHapgsResponse :: ListHapgsResponse

ListHapgsResponse constructor.

The fields accessible through corresponding lenses are:

lhr1HapgList :: [Text]

lhr1NextToken :: Maybe Text

Response lenses

lhr1HapgList :: Lens' ListHapgsResponse [Text]

The list of high-availability partition groups.

lhr1NextToken :: Lens' ListHapgsResponse (Maybe Text)

If not null, more results are available. Pass this value to ListHapgs to retrieve the next set of items.