Network.AWS.CloudHSM.ListHsms

Description

Retrieves the identifiers of all of the HSMs provisioned for the current customer.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHsms to retrieve the next set of items.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_ListHsms.html

Request

data ListHsms

Request constructor

listHsms :: ListHsms

ListHsms constructor.

The fields accessible through corresponding lenses are:

lh1NextToken :: Maybe Text

Request lenses

lh1NextToken :: Lens' ListHsms (Maybe Text)

The NextToken value from a previous call to ListHsms. Pass null if this is the first call.

Response

data ListHsmsResponse

Response constructor

listHsmsResponse :: ListHsmsResponse

ListHsmsResponse constructor.

The fields accessible through corresponding lenses are:

lhrHsmList :: [Text]

lhrNextToken :: Maybe Text

Response lenses

lhrHsmList :: Lens' ListHsmsResponse [Text]

The list of ARNs that identify the HSMs.

lhrNextToken :: Lens' ListHsmsResponse (Maybe Text)

If not null, more results are available. Pass this value to ListHsms to retrieve the next set of items.