Network.AWS.CloudHSM.ListLunaClients

Description

Lists all of the clients.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListLunaClients to retrieve the next set of items.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_ListLunaClients.html

Request

data ListLunaClients

Request constructor

listLunaClients :: ListLunaClients

ListLunaClients constructor.

The fields accessible through corresponding lenses are:

llcNextToken :: Maybe Text

Request lenses

llcNextToken :: Lens' ListLunaClients (Maybe Text)

The NextToken value from a previous call to ListLunaClients. Pass null if this is the first call.

Response

data ListLunaClientsResponse

Response constructor

listLunaClientsResponse :: ListLunaClientsResponse

ListLunaClientsResponse constructor.

The fields accessible through corresponding lenses are:

llcrClientList :: [Text]

llcrNextToken :: Maybe Text

Response lenses

llcrClientList :: Lens' ListLunaClientsResponse [Text]

The list of clients.

llcrNextToken :: Lens' ListLunaClientsResponse (Maybe Text)

If not null, more results are available. Pass this to ListLunaClients to retrieve the next set of items.