Network.AWS.CloudHSM.ModifyHSM

Description

Modifies an HSM.

Important: This operation can result in the HSM being offline for up to 15 minutes while the AWS CloudHSM service is reconfigured. If you are modifying a production HSM, you should ensure that your AWS CloudHSM service is configured for high availability, and consider executing this operation during a maintenance window.

Creating a Request

modifyHSM

Arguments

:: Text
mhHSMARN

-> ModifyHSM

Creates a value of ModifyHSM with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mhIAMRoleARN - The new IAM role ARN.

mhSubnetId - The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

mhSyslogIP - The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

mhExternalId - The new external ID.

mhEniIP - The new IP address for the elastic network interface (ENI) attached to the HSM. If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

mhHSMARN - The ARN of the HSM to modify.

data ModifyHSM

Contains the inputs for the ModifyHsm operation.

See: modifyHSM smart constructor.

Request Lenses

mhIAMRoleARN :: Lens' ModifyHSM (Maybe Text)

The new IAM role ARN.

mhSubnetId :: Lens' ModifyHSM (Maybe Text)

The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

mhSyslogIP :: Lens' ModifyHSM (Maybe Text)

The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

mhExternalId :: Lens' ModifyHSM (Maybe Text)

The new external ID.

mhEniIP :: Lens' ModifyHSM (Maybe Text)

The new IP address for the elastic network interface (ENI) attached to the HSM. If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

mhHSMARN :: Lens' ModifyHSM Text

The ARN of the HSM to modify.

Destructuring the Response

modifyHSMResponse

Arguments

:: Int
mhsmrsResponseStatus

-> ModifyHSMResponse

Creates a value of ModifyHSMResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mhsmrsHSMARN - The ARN of the HSM.

mhsmrsResponseStatus - -- | The response status code.

data ModifyHSMResponse

Contains the output of the ModifyHsm operation.

See: modifyHSMResponse smart constructor.

Response Lenses

mhsmrsHSMARN :: Lens' ModifyHSMResponse (Maybe Text)

The ARN of the HSM.

mhsmrsResponseStatus :: Lens' ModifyHSMResponse Int

- | The response status code.