Network.AWS.CloudHSM.ModifyHsm

Description

Modifies an HSM.

http://docs.aws.amazon.com/cloudhsm/latest/dg/API_ModifyHsm.html

Request

data ModifyHsm

Request constructor

modifyHsm

Arguments

:: Text
mhHsmArn

-> ModifyHsm

ModifyHsm constructor.

The fields accessible through corresponding lenses are:

mhEniIp :: Maybe Text

mhExternalId :: Maybe Text

mhHsmArn :: Text

mhIamRoleArn :: Maybe Text

mhSubnetId :: Maybe Text

mhSyslogIp :: Maybe Text

Request lenses

mhEniIp :: Lens' ModifyHsm (Maybe Text)

The new IP address for the elastic network interface attached to the HSM.

mhExternalId :: Lens' ModifyHsm (Maybe Text)

The new external ID.

mhHsmArn :: Lens' ModifyHsm Text

The ARN of the HSM to modify.

mhIamRoleArn :: Lens' ModifyHsm (Maybe Text)

The new IAM role ARN.

mhSubnetId :: Lens' ModifyHsm (Maybe Text)

The new identifier of the subnet that the HSM is in.

mhSyslogIp :: Lens' ModifyHsm (Maybe Text)

The new IP address for the syslog monitoring server.

Response

data ModifyHsmResponse

Response constructor

modifyHsmResponse :: ModifyHsmResponse

ModifyHsmResponse constructor.

The fields accessible through corresponding lenses are:

mhrHsmArn :: Maybe Text

Response lenses

mhrHsmArn :: Lens' ModifyHsmResponse (Maybe Text)

The ARN of the HSM.