Network.AWS.CloudHSM.ModifyLunaClient

Description

Modifies the certificate used by the client.

This action can potentially start a workflow to install the new certificate on the client's HSMs.

Creating a Request

modifyLunaClient

Arguments

:: Text
mlcClientARN

-> Text
mlcCertificate

-> ModifyLunaClient

Creates a value of ModifyLunaClient with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlcClientARN - The ARN of the client.

mlcCertificate - The new certificate for the client.

data ModifyLunaClient

See: modifyLunaClient smart constructor.

Request Lenses

mlcClientARN :: Lens' ModifyLunaClient Text

The ARN of the client.

mlcCertificate :: Lens' ModifyLunaClient Text

The new certificate for the client.

Destructuring the Response

modifyLunaClientResponse

Arguments

:: Int
mlcrsResponseStatus

-> ModifyLunaClientResponse

Creates a value of ModifyLunaClientResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlcrsClientARN - The ARN of the client.

mlcrsResponseStatus - -- | The response status code.

data ModifyLunaClientResponse

See: modifyLunaClientResponse smart constructor.

Response Lenses

mlcrsClientARN :: Lens' ModifyLunaClientResponse (Maybe Text)

The ARN of the client.

mlcrsResponseStatus :: Lens' ModifyLunaClientResponse Int

- | The response status code.