Network.AWS.CloudHSM

Description

AWS CloudHSM Service

Service Configuration

cloudHSM :: Service

API version 2014-05-30 of the Amazon CloudHSM SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudHSM.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that one or more of the request parameters are not valid.

CloudHSMServiceException

_CloudHSMServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an exception occurred in the AWS CloudHSM service.

CloudHSMInternalException

_CloudHSMInternalException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an internal error occurred.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteHAPG

module Network.AWS.CloudHSM.DeleteHAPG

ListHAPGs

module Network.AWS.CloudHSM.ListHAPGs

ModifyLunaClient

module Network.AWS.CloudHSM.ModifyLunaClient

ListHSMs

module Network.AWS.CloudHSM.ListHSMs

DescribeLunaClient

module Network.AWS.CloudHSM.DescribeLunaClient

ListTagsForResource

module Network.AWS.CloudHSM.ListTagsForResource

CreateHAPG

module Network.AWS.CloudHSM.CreateHAPG

CreateHSM

module Network.AWS.CloudHSM.CreateHSM

RemoveTagsFromResource

module Network.AWS.CloudHSM.RemoveTagsFromResource

DescribeHAPG

module Network.AWS.CloudHSM.DescribeHAPG

CreateLunaClient

module Network.AWS.CloudHSM.CreateLunaClient

ListLunaClients

module Network.AWS.CloudHSM.ListLunaClients

AddTagsToResource

module Network.AWS.CloudHSM.AddTagsToResource

GetConfig

module Network.AWS.CloudHSM.GetConfig

DeleteHSM

module Network.AWS.CloudHSM.DeleteHSM

DescribeHSM

module Network.AWS.CloudHSM.DescribeHSM

ModifyHAPG

module Network.AWS.CloudHSM.ModifyHAPG

DeleteLunaClient

module Network.AWS.CloudHSM.DeleteLunaClient

ModifyHSM

module Network.AWS.CloudHSM.ModifyHSM

ListAvailableZones

module Network.AWS.CloudHSM.ListAvailableZones

Types

ClientVersion

data ClientVersion

Constructors

VD5_1

VD5_3

CloudHSMObjectState

data CloudHSMObjectState

Constructors

Degraded

Ready

Updating

HSMStatus

data HSMStatus

Constructors

HSDegraded

HSPending

HSRunning

HSSuspended

HSTerminated

HSTerminating

HSUpdating

SubscriptionType

data SubscriptionType

Specifies the type of subscription for the HSM.

PRODUCTION - The HSM is being used in a production environment. * TRIAL - The HSM is being used in a product trial.

Constructors

Production

Tag

data Tag

A key-value pair that identifies or specifies metadata about an AWS CloudHSM resource.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - The key of the tag.

tagValue - The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

tagValue :: Lens' Tag Text

The value of the tag.