Network.AWS.CloudHSM.ListLunaClients

Description

Lists all of the clients.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListLunaClients to retrieve the next set of items.

Creating a Request

listLunaClients :: ListLunaClients

Creates a value of ListLunaClients with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

llcNextToken - The NextToken value from a previous call to ListLunaClients . Pass null if this is the first call.

data ListLunaClients

See: listLunaClients smart constructor.

Request Lenses

llcNextToken :: Lens' ListLunaClients (Maybe Text)

The NextToken value from a previous call to ListLunaClients . Pass null if this is the first call.

Destructuring the Response

listLunaClientsResponse

Arguments

:: Int
llcrsResponseStatus

-> ListLunaClientsResponse

Creates a value of ListLunaClientsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

llcrsNextToken - If not null, more results are available. Pass this to ListLunaClients to retrieve the next set of items.

llcrsResponseStatus - -- | The response status code.

llcrsClientList - The list of clients.

data ListLunaClientsResponse

See: listLunaClientsResponse smart constructor.

Response Lenses

llcrsNextToken :: Lens' ListLunaClientsResponse (Maybe Text)

If not null, more results are available. Pass this to ListLunaClients to retrieve the next set of items.

llcrsResponseStatus :: Lens' ListLunaClientsResponse Int

- | The response status code.

llcrsClientList :: Lens' ListLunaClientsResponse [Text]

The list of clients.